requirements.txt
PyYAML==3.11
docopt==0.6.1
termcolor==1.1.0
httplib2==0.9

